# backend/internal/database/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package database

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type ApiKey struct {
	ID         pgtype.UUID        `json:"id"`
	ProjectID  pgtype.UUID        `json:"project_id"`
	Name       string             `json:"name"`
	KeyHash    string             `json:"key_hash"`
	KeySuffix  string             `json:"key_suffix"`
	CreatedAt  pgtype.Timestamptz `json:"created_at"`
	LastUsedAt pgtype.Timestamptz `json:"last_used_at"`
}

type Asset struct {
	ID              pgtype.UUID `json:"id"`
	UpdateID        pgtype.UUID `json:"update_id"`
	Platform        string      `json:"platform"`
	FileName        string      `json:"file_name"`
	MimeType        string      `json:"mime_type"`
	Key             string      `json:"key"`
	Url             string      `json:"url"`
	FileHash        string      `json:"file_hash"`
	Hash            string      `json:"hash"`
	StorageProvider string      `json:"storage_provider"`
}

type DownloadEvent struct {
	ID         int64              `json:"id"`
	UpdateID   pgtype.UUID        `json:"update_id"`
	ProjectID  pgtype.UUID        `json:"project_id"`
	Timestamp  pgtype.Timestamptz `json:"timestamp"`
	DeviceHash string             `json:"device_hash"`
	Platform   string             `json:"platform"`
	Channel    string             `json:"channel"`
}

type Project struct {
	ID          pgtype.UUID        `json:"id"`
	Slug        string             `json:"slug"`
	Name        string             `json:"name"`
	Description string             `json:"description"`
	CreatedAt   pgtype.Timestamptz `json:"created_at"`
}

type Setting struct {
	Key       string             `json:"key"`
	Value     string             `json:"value"`
	UpdatedAt pgtype.Timestamptz `json:"updated_at"`
}

type Update struct {
	ID                pgtype.UUID        `json:"id"`
	ProjectID         pgtype.UUID        `json:"project_id"`
	RuntimeVersion    string             `json:"runtime_version"`
	Channel           string             `json:"channel"`
	RolloutPercentage int32              `json:"rollout_percentage"`
	Platform          string             `json:"platform"`
	IsActive          bool               `json:"is_active"`
	IsRollback        bool               `json:"is_rollback"`
	Message           pgtype.Text        `json:"message"`
	CreatedAt         pgtype.Timestamptz `json:"created_at"`
	ExpoConfig        []byte             `json:"expo_config"`
}